Refactor admin components
