+++ pages/_app.js
@@ -142,7 +142,7 @@ class Layout extends React.Component {
         <Head>
           <link rel='stylesheet' href='https://unpkg.com/leaflet@1.4.0/dist/leaflet.css'
             integrity='sha512-puBpdR0798OZvTTbP4A8Ix/l+A4dHDD0DGqYW6RQ+9jxkRFclaxxQb/SJAWZfWAkuyeQUytO7+7N4QKrDh+drA=='
-            crossorigin='' />
+            crossOrigin='' />
         </Head>
         <header className='header-nav'>
           <div className='row'>
